Parse the key id of GPG keys and store them in the database
I'm sending over a patch that parses the key id of GPG keys and stores them in
a newly added column, so that it's parsed only when it's uploaded (any
existing DBs will have to be recreated). It assumes that gpg2 is in $PATH.

The patch also has a couple of other minor changes:

- allow non @debian.org emails in debug mode
- add a bit of padding in the HTML table that lists uploaded packages
 (previously some cells looked crammed together)

Sorry for lumping it all together, but it's kind of trivial anyway.

Also, I found that a decorator throws a KeyError exception when a given
password is less than 6 chars (the min requirement), but I wasn't able to fix
that (I'm kind of inexperienced with decorators).